Centralize tile behavior
